Extended TrumpTweetDataset
- Added shuffle method.
- Added years parameter.
- Added chain tweet filter.
- Added export code.

+++ Datasets.py
@@ -1,11 +1,11 @@
+import json
+import os
 import re
 from urllib import request
-import os
 
 import numpy as np
 from torch.utils.data.dataset import Dataset
-import tempfile
-import json
+from tqdm import trange
 
 at_regex = re.compile(r'(@\w{1,15})')
 hash_regex = re.compile(r'(#[\w_\d]+)')
@@ -15,15 +15,16 @@ url_regex = re.compile(
 )
 clean_regex = re.compile(r'[^\w\-_!?:.,;()\d\'"\s#@]')
 
+
 # File example (url): http://trumptwitterarchive.com/data/realdonaldtrump/2019.json
 
 # TODO: Make Iterable Over Persons (other than trump)?
 class TrumpTweetDataset(Dataset):
-    def __init__(self, save_dir=None, download=True):
+    def __init__(self, save_dir=None, download=True, years=range(2009, 2020)):
         super(TrumpTweetDataset, self).__init__()
         file_list = []
 
-        for year in range(2015, 2019):
+        for year in years:
             file_list.append(str(year) + ".json")
 
         if save_dir is None:
@@ -33,8 +34,10 @@ class TrumpTweetDataset(Dataset):
             url_root = 'http://trumptwitterarchive.com/data/realdonaldtrump/'
 
             for file in file_list:
-                print(f'Downloading {file} to {save_dir}..')
                 file_path = os.path.join(save_dir, file)
+                if os.path.exists(file_path):
+                    continue
+                print(f'Downloading {file} to {save_dir}..')
                 request.urlretrieve(url_root + file, file_path)
 
         # join tweets from all selected years in a single list
@@ -46,7 +49,14 @@ class TrumpTweetDataset(Dataset):
         print(f'loaded {len(self.tweets)} raw tweets')
 
         # remove re-tweets
-        self.tweets = list(filter(lambda tw: not tw.get('is_retweet', True), self.tweets))
+        self.tweets = list(
+            filter(lambda tw: not tw.get('is_retweet', True),
+                   self.tweets))
+
+        # remove chain tweets
+        self.tweets = list(
+            filter(lambda tw: not (tw.get('text', "").endswith("..") or tw.get('text', "").startswith("..")),
+                   self.tweets))
         print(f'loaded {len(self.tweets)} clean tweets')
 
     def __getitem__(self, item):
@@ -60,7 +70,7 @@ class TrumpTweetDataset(Dataset):
         text = clean_regex.sub('', text)
 
         # replace @-handles with masks
-        if False: # TODO: parametrize replacers
+        if False:  # TODO: parametrize replacers
             text = at_regex.sub('[TW-AT]', text)
 
         # TODO: Create hash regex replacer
@@ -73,6 +83,22 @@ class TrumpTweetDataset(Dataset):
     def __len__(self):
         return len(self.tweets)
 
+    def shuffle(self):
+        np.random.shuffle(self.tweets)
 
-data = TrumpTweetDataset(download=True)
-print(data[0])
+
+for year in (2009, 2015, 2016):
+    data = TrumpTweetDataset(download=True, years=range(year, 2020))
+    data.shuffle()
+    prefix = f'trump_tweets_{year}-2019'
+    with open(prefix + '.train', 'w', encoding='utf-8') as train, \
+            open(prefix + '.dev', 'w', encoding='utf-8') as dev, \
+            open(prefix + '.test', 'w', encoding='utf-8') as test:
+        train_sample = int(len(data) * 0.8)
+        dev_sample = int(len(data) * 0.9)
+        for i in trange(train_sample):
+            train.write(data[i] + "\n")
+        for i in trange(train_sample, dev_sample):
+            dev.write(data[i] + "\n")
+        for i in trange(dev_sample, len(data)):
+            test.write(data[i] + "\n")
